test(write): colocate fixed-bucket builder coverage
